+++ src/Axle.Client/Models/UserSessionResponse.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) Lykke Corp.
+// See the LICENSE file in the project root for more information.
+
+namespace Axle.Client.Models
+{
+    public class UserSessionResponse
+    {
+        public int UserSessionId { get; set; }
+    }
+}