+++ src/Axle.Persistence/ISessionRepository.cs
@@ -5,10 +5,19 @@ namespace Axle.Persistence
 {
     using System.Collections.Generic;
 
-    public interface ISessionRepository : IRepository<int, Session>
+    public interface ISessionRepository
     {
-        Session GetByUser(string userId);
+        void Add(Session entity);
+
+#pragma warning disable CA1716 // Identifiers should not match keywords
+        Session Get(int sessionId);
+
+        Session GetByUser(string userName);
+
+        void Remove(int sessionId, string userName);
 
         void RefreshSessionTimeouts(IEnumerable<Session> sessions);
+
+        IEnumerable<Session> GetExpiredSessions();
     }
 }