+++ src/Axle/Program.cs
@@ -23,6 +23,7 @@ namespace Axle
             ("Api-Name",                    "API_NAME",                    null),
             ("Api-Secret",                  "API_SECRET",                  null),
             ("ConnectionStrings:Redis",     "REDIS_CONNECTIONSTRING",      null),
+            ("ConnectionStrings:RabbitMq",  "RABBITMQ_CONNECTIONSTRING",   null),
             ("Require-Https",               "REQUIRE_HTTPS",               "true"),
             ("Swagger-Client-Id",           "SWAGGER_CLIENT_ID",           "axle_api_swagger"),
             ("Validate-Issuer-Name",        "VALIDATE_ISSUER_NAME",        "false")