+++ src/Axle/Services/SessionLifecycleService.cs
@@ -8,6 +8,7 @@ namespace Axle.Services
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
+    using Axle.Contracts;
     using Axle.Dto;
     using Axle.Persistence;
     using Microsoft.Extensions.Logging;
@@ -20,6 +21,7 @@ namespace Axle.Services
         private readonly ISessionRepository sessionRepository;
         private readonly ITokenRevocationService tokenRevocationService;
         private readonly INotificationService notificationService;
+        private readonly IActivityService activityService;
         private readonly ILogger<SessionLifecycleService> logger;
         private readonly TimeSpan sessionTimeout;
 
@@ -31,12 +33,14 @@ namespace Axle.Services
             ISessionRepository sessionRepository,
             ITokenRevocationService tokenRevocationService,
             INotificationService notificationService,
+            IActivityService activityService,
             ILogger<SessionLifecycleService> logger,
             TimeSpan sessionTimeout)
         {
             this.sessionRepository = sessionRepository;
             this.tokenRevocationService = tokenRevocationService;
             this.notificationService = notificationService;
+            this.activityService = activityService;
             this.logger = logger;
             this.sessionTimeout = sessionTimeout;
 
@@ -67,7 +71,13 @@ namespace Axle.Services
             }
         }
 
-        public async Task OpenConnection(string connectionId, string userId, string clientId, string accessToken)
+        public async Task OpenConnection(
+            string connectionId,
+            string userName,
+            string accountId,
+            string clientId,
+            string accessToken,
+            bool isSupportUser)
         {
             Session userInfo;
 
@@ -75,7 +85,7 @@ namespace Axle.Services
 
             try
             {
-                userInfo = this.sessionRepository.GetByUser(userId);
+                userInfo = this.sessionRepository.GetByUser(userName);
 
                 if (userInfo != null && userInfo.AccessToken == accessToken)
                 {
@@ -92,14 +102,19 @@ namespace Axle.Services
                 }
                 while (this.sessionRepository.Get(sessionId) != null);
 
-                var newState = new Session(userId, sessionId, accessToken, clientId);
+                var newSession = new Session(userName, sessionId, accountId, accessToken, clientId, isSupportUser);
 
-                this.sessionRepository.Add(sessionId, newState);
-                this.connectionSessionMap.TryAdd(connectionId, newState);
+                this.sessionRepository.Add(newSession);
+                this.connectionSessionMap.TryAdd(connectionId, newSession);
+
+                if (!newSession.IsSupportUser)
+                {
+                    await this.activityService.PublishActivity(newSession, SessionActivityType.Login);
+                }
 
                 if (userInfo != null)
                 {
-                    await this.TerminateSession(userInfo);
+                    await this.TerminateSession(userInfo, SessionActivityType.DifferentDeviceTermination);
                 }
             }
             finally
@@ -108,28 +123,30 @@ namespace Axle.Services
             }
         }
 
-        public async Task<TerminateSessionResponse> TerminateSession(string userId)
+        public async Task<TerminateSessionResponse> TerminateSession(
+            string userName,
+            SessionActivityType reason = SessionActivityType.ManualTermination)
         {
             await this.slimLock.WaitAsync();
 
             try
             {
-                var userInfo = this.sessionRepository.GetByUser(userId);
+                var userInfo = this.sessionRepository.GetByUser(userName);
 
                 if (userInfo == null)
                 {
                     return new TerminateSessionResponse
                     {
                         Status = TerminateSessionStatus.NotFound,
-                        ErrorMessage = $"No session found for the user: [{userId}]"
+                        ErrorMessage = $"No session found for the user: [{userName}]"
                     };
                 }
 
-                this.logger.LogInformation($"Terminating session: [{userInfo.SessionId}] for user: [{userId}]");
+                this.logger.LogInformation($"Terminating session: [{userInfo.SessionId}] for user: [{userName}]");
 
-                await this.TerminateSession(userInfo);
+                await this.TerminateSession(userInfo, reason);
 
-                this.logger.LogInformation($"Successfully terminated session: [{userInfo.SessionId}] for user: [{userId}]");
+                this.logger.LogInformation($"Successfully terminated session: [{userInfo.SessionId}] for user: [{userName}]");
 
                 return new TerminateSessionResponse
                 {
@@ -139,7 +156,7 @@ namespace Axle.Services
             }
             catch (Exception error)
             {
-                this.logger.LogError(error, $"An unexpected error occurred while terminating session for user [{userId}]");
+                this.logger.LogError(error, $"An unexpected error occurred while terminating session for user [{userName}]");
 
                 return new TerminateSessionResponse
                 {
@@ -153,12 +170,18 @@ namespace Axle.Services
             }
         }
 
-        public async Task TerminateSession(Session userInfo)
+        public async Task TerminateSession(Session userInfo, SessionActivityType reason)
         {
-            this.sessionRepository.Remove(userInfo.SessionId);
+            this.sessionRepository.Remove(userInfo.SessionId, userInfo.UserName);
             await this.tokenRevocationService.RevokeAccessToken(userInfo.AccessToken, userInfo.ClientId);
 
             this.notificationService.PublishSessionTermination(userInfo.SessionId);
+
+            // Support user activities are not required currently
+            if (!userInfo.IsSupportUser)
+            {
+                await this.activityService.PublishActivity(userInfo, reason);
+            }
         }
 
         public void Dispose()
@@ -180,6 +203,14 @@ namespace Axle.Services
                     try
                     {
                         this.sessionRepository.RefreshSessionTimeouts(this.connectionSessionMap.Values);
+
+                        var sessionsToTerminate = this.sessionRepository.GetExpiredSessions();
+
+                        foreach (var session in sessionsToTerminate)
+                        {
+                            await this.TerminateSession(session, SessionActivityType.TimeOut);
+                            this.logger.LogInformation($"Successfully timed out session: [{session.SessionId}] for user: [{session.UserName}]");
+                        }
                     }
                     finally
                     {